+++ forge-1-16/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+# [1.16.5-1.0.0.beta2] - 21/12/2023
+
+### Added Blocks. Ore generation, Recipes, etc. will be added in the coming updates.
+
 # [1.16.5-1.0.0.beta1] - 20/12/2023
 
 ### Introduced 1.16.5 support for the mod. Currently only has items. Ore generation, blocks, etc. will be added later.
