+++ forge-1-16/gradle.properties
@@ -2,7 +2,7 @@ kotlin.code.style=official
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 
-modVersion = 1.16.5-1.0.0.beta1
+modVersion = 1.16.5-1.0.0.beta2
 mcVersion = 1.16.5
 projectId = 480779
 modId = temporalsmith
