+++ forge-1-16/src/main/java/io/github/realyusufismail/temporalsmith/lit/RainbowLitBlock.java
@@ -16,15 +16,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */ 
-package blocks.lit;
+package io.github.realyusufismail.temporalsmith.lit;
 
+import java.util.Arrays;
 import java.util.Collections;
 
-public class RubyLitBlock extends LITBlock {
-  public RubyLitBlock() {
+public class RainbowLitBlock extends LITBlock {
+  public RainbowLitBlock() {
     super(
-        Collections.singletonList(LitBlockParticleColour.RUBY_PARTICLE_COLOR),
+        Collections.unmodifiableList(
+            Arrays.asList(
+                LitBlockParticleColour.RAINBOW_PARTICLE_COLOR_RED,
+                LitBlockParticleColour.RAINBOW_PARTICLE_COLOR_GREEN,
+                LitBlockParticleColour.RAINBOW_PARTICLE_COLOR_BLUE,
+                LitBlockParticleColour.RAINBOW_PARTICLE_COLOR_PURPLE)),
         11,
-        RubyLitBlock.class);
+        RainbowLitBlock.class);
   }
 }
