+++ forge-1-16/src/main/resources/META-INF/mods.toml
@@ -4,7 +4,7 @@ license = "Apache 2.0" #mandatory
 issueTrackerURL = "https://github.com/RealYusufIsmail-Mc-Mods/TemporalSmith"
 [[mods]]
 modId = "temporalsmith"
-version = "1.16.5-1.0.0.beta1"
+version = "1.16.5-1.0.0.beta2"
 displayName = "TemporalSmith: Aetheric Arsenal Expansion"
 logoFile = "logo.png"
 updateJSONURL = "https://forge.curseupdate.com/480779/temporalsmith"
